Changed to executing MP2K commands by event
Now it executes MP2K commands by event, instead of re-reading from ROM, as the ROM should only need to be read once

Many changes have also been made as well, including the beginning of changes to the Track Editor and its ability to read and change params, more to come!

Many libraries have also been updated, and needed to change stuff to reflect the changes

+++ VG Music Studio - Core/Config.cs
@@ -8,7 +8,7 @@ namespace Kermalis.VGMusicStudio.Core;
 public abstract class Config : IDisposable
 {
     public virtual byte[]? ROM { get; }
-    public int[]? SongTableOffset { get; internal set; }
+    public long[]? SongTableOffset { get; internal set; }
     public virtual HSLColor[]? Colors { get; protected set; }
 
     public readonly struct Song
